Separating FortranString into multiple files---per Style Guide.
