+++ src/njoy21/input/argument/parser/FortranString/src/isQuote.hpp
@@ -1,0 +1,8 @@
+bool isQuote( char c ){
+  if ( ( c == '\'' ) or ( c == '*' ) or ( c == '"' ) ){
+    quote = c;
+    return true;
+  }
+  return false;
+}
+
